check files exist

+++ install_hdf5.sh
@@ -5,13 +5,18 @@
 # sudo apt install -q -y libhdf5-dev
 
 # szip dependency
-wget https://support.hdfgroup.org/ftp/lib-external/szip/2.1.1/src/szip-2.1.1.tar.gz
-tar -xvzf szip-2.1.1.tar.gz
-rm -f szip-2.1.1.tar.gz 
+if [[ ! -d szip-2.1.1 ]]; then
+    wget https://support.hdfgroup.org/ftp/lib-external/szip/2.1.1/src/szip-2.1.1.tar.gz
+    tar -xvzf szip-2.1.1.tar.gz
+    rm -f szip-2.1.1.tar.gz 
+fi
 cd szip-2.1.1
 mkdir /home/ubuntu/szip
 ./configure --prefix=/home/ubuntu/szip
-make
+make >> /dev/null
+if [[! $?]] 
+    
+fi
 make check
 make install
 cd ..
@@ -26,6 +31,16 @@ rm -rf szip-2.1.1
 wget https://support.hdfgroup.org/ftp/HDF5/current/src/hdf5-1.10.1.tar
 tar -xvf hdf5-1.10.1.tar
 rm -f hdf5-1.10.1.tar
+
+export CC=icc
+export F9X=ifort
+export CXX=icpc
+
+cd hdf5-1.10.1
+./configure --prefix=/usr/local/hdf5-1.10.1 --enable-fortran --enable-cxx
+make
+make check
+make install
 # https://github.com/HDFGroup/build_hdf5
 # https://support.hdfgroup.org/ftp/HDF5/current/src/unpacked/release_docs/INSTALL_parallel
 # https://support.hdfgroup.org/HDF5/release/obtainsrc.html
